[Delete] Finished CRUD Migration (#11)
